+++ CHANGELOG.md
@@ -85,6 +85,7 @@
 - PIM-10705: Revert Imagick time limit because it is stateful
 - PIM-10698: Fix PHP Warning: Undefined array key "locale" in Tailored Import
 - PIM-10693: Improve the speed up of the process stop time on the rules
+- PIM-10692: Fix wrong DQI assessments on select attribute
 
 ## Improvements